Restore Unity 6 compatibility, fix escort duty, add dodge assist, launch pads, smart beacon
Game update June 2026 moved DRG Survivor to Unity 6 (IL2CPP metadata v39),
breaking BepInEx entirely. This release:

- Requires BepInEx bleeding edge be.785 (metadata v39 support)
- Adds tools/InteropFixer: repairs interop assemblies corrupted by the
  Il2CppInterop generator on Unity 6 games (orphan nested TypeDefs causing
  "Duplicate type with name '<>O'" BadImageFormatException)
- Adapts to game API changes (long/double stat types, moved unlock ranks,
  removed endless-mode end screen fields, UIElementsModule reference)
- Fixes escort duty: TNT phase no longer triggers at mission start
  (detonators are map blocks present from level generation); phase
  detection now polls real game state with patches as accelerators
- Fixes audio systems dying after Retry (13 components had a broken
  destroyed-GameController check; now all use State-read validation)
- Adds Dreadnought dodge assistance: charge-path corridor warnings with
  escape direction, spike damage-radius alerts, fireball move nudge
- Adds launch pad 8-bit "boing" beacon with name-based scanning and
  runtime diagnostics to discover pad prefab names
- Adds optional Smart Beacon: priority-scored collectible tracking
  (Hades 2 style), healing prioritized when hurt
- Uses Bobby.MAX_FUEL instead of assuming 100 for fuel warnings
- Adds localization for all new strings in 22 languages
- Version 0.10.0

+++ drgAccess/Patches/PickupAnnouncementPatches.cs
@@ -28,7 +28,7 @@ public static class PickupAnnouncementPatches
             var healType = args.Type;
             if (healType != EHealType.HEAL) return;
 
-            int healed = args.ActualHeal;
+            long healed = args.ActualHeal;
             if (healed <= 0) return;
 
             ScreenReader.Say(ModLocalization.Get("pickup_healed", healed));
